// projects/pos/src/app/shared/enums/columns/product-columns.enum.ts
export enum ProductColumnLabel {
  NAME = 'Name',
  SLUG = 'Slug',
  SKU = 'SKU',
  TYPE = 'Type',
  IS_NEW = 'New',
  QUANTITY = 'Quantity',
  COST = 'Cost',
  MRP = 'MRP',
  PRICE = 'Price',
  SIZE = 'Size',
  COLOR = 'Color',
  STATUS = 'Status',
}

export enum ProductColumn {
  NAME = 'name',
  SLUG = 'slug',
  SKU = 'sku',
  TYPE = 'type',
  IS_NEW = 'is_new',
  QUANTITY = 'quantity',
  COST = 'cost',
  MRP = 'mrp',
  PRICE = 'price',
  SIZE = 'size',
  COLOR = 'color',
  STATUS = 'status',
}